ci(build): speed up multi-platform builds, and add arm/v6 and arm/v7 support (#5)
This is a bit of "hacky" solution to for speedy builds. It essentially
uses goreleaser to build a suite of multi-platform binaries on the host
platform, then exposes them to the Docker build process with a
Dockerfile which will grab the correct binary depending on the target
platform that the build process is running on.